Configure the parser via a TOML file.
The number of command-line arguments was getting a bit out of hand. Use one
configuration file for the parser that can be used during training and
evaluation.

File names specified in the configuration are currently relative to the
current working directory. It may be worth considering to read files
relative to the configuration file directory.