Add FreezableLookup which replaces several lookup tables.
This change introduces a single lookup table that replaces:

- TransitionLookup
- LookupTable
- MutableLookupTable

StoredLookupTable is simplified, but still exists as a wrapper around
FreezableLookup to simplify serialization.

FreezableLookup implements three traits:

- LookupValue: defines the methods to look up an identifier for a value
  and vice versa.
- LookupLen: defines the `len` method.
- Lookup: inherits LookupValue and LookupLen and is implemented for any
  type that implements both traits.

In my first iteration, I had a single lookup trait. However, this
sometimes led to type inference problems when the `len` method was used.
Since `len` does not use the borrowed type in its signature, the type
parameter for Lookup<B> could not be inferred. So, instead, we make the
`len` method part of a trait without type parameters.

Fixes #19.