Use rust2vec with finalfusion word representations
With this change, dpar can use word representations with subword units. This
also makes the character/morphology layer obsolete. The morphology layer is
therefore removed.

Since finalfusion embeddings can be computed on the fly, we cannot use the
previous method of passing the word indices to Tensorflow. This change adds
an additional layer to the input that contains the concatenation of all
embeddings.

Indices can still be used for all layers. If a layer is encoded using
indices, then the 'traditional' lookup layer is used.